friday: fixed flex-items. converted blog-items to grid. focused on making responsive content. added media queries. Saturday: focused on topnav. fixed some items not behaving as intended.
ps: forgot to push on friday

have been struggling with responsive content like fonts. especially on blog items. seems to be working now. not happy about the menu hack on mobile and tablet, but as it is right now, cant do much else.

i have been struggling with my confidence with my design. i feel inferior to my classmates, but i have to work past it and focus on coding. the design i have made .